bot/requirements.txt
nest_asyncio
asyncio
pyTelegramBotAPI
python-socketio
python-decouple
aiohttp